tests/submission/W3C/domhtml/area03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Area</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/editing.html#dom-tabindex" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<p>
<map id="mapid" name="mapid">
<area accesskey="a" alt="Domain" coords="0,2,45,45" href="./files/dletter.html" shape="rect" tabindex="10" title="Domain"></area>
</map>
</p>


<script>
var domtest = async_test("area03");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vtabindex;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("area");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vtabindex = testNode.tabIndex;

      assert_equals(vtabindex,10);
       
    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
